update training sdk config to combine config and status file into one unified_config file
