+++ src/sagemaker/hyperpod/training/__init__.py
@@ -1,7 +1,4 @@
-from sagemaker.hyperpod.training.config.hyperpod_pytorch_job_config import *
-from sagemaker.hyperpod.training.config.hyperpod_pytorch_job_status import (
-    HyperPodPytorchJobStatus,
-)
+from sagemaker.hyperpod.training.config.hyperpod_pytorch_job_unified_config import *
 from sagemaker.hyperpod.training.hyperpod_pytorch_job import (
     HyperPodPytorchJob,
     _load_hp_job,
